Fixes int for applications ids

+++ include/NetworkInterface.h
@@ -550,8 +550,8 @@ public:
   inline char *get_ndpi_proto_name(u_int id) {
     return (ndpi_get_proto_name(get_ndpi_struct(), id));
   };
-  inline int get_ndpi_proto_id(char *proto) {
-    int _proto = ndpi_get_protocol_id(get_ndpi_struct(), proto);
+  inline u_int16_t get_ndpi_proto_id(char *proto) {
+    u_int16_t _proto = ndpi_get_protocol_id(get_ndpi_struct(), proto);
     return ndpi_map_ndpi_id_to_user_proto_id(get_ndpi_struct(), _proto);
   };
   inline int get_ndpi_category_id(char *cat) {
@@ -1289,7 +1289,7 @@ public:
   inline ndpi_protocol_category_t get_ndpi_proto_category(ndpi_protocol proto) {
     return (ndpi_get_proto_category(get_ndpi_struct(), proto));
   };
-  ndpi_protocol_category_t get_ndpi_proto_category(u_int protoid);
+  ndpi_protocol_category_t get_ndpi_proto_category(u_int16_t protoid);
   void setnDPIProtocolCategory(u_int16_t protoId,
                                ndpi_protocol_category_t protoCategory);
   bool nDPILoadIPCategory(char *what, ndpi_protocol_category_t id,
